Adjust note durations to time signature
Different time signatures use different note values as the beat. When
you render a midi file, you have to convert the notes to midi durations
that correspond with the beat so that the tempo sounds correct. By
using the quarter note as the default beat, you can scale your note
durations for any time signature.