Refactor search algorithm
The process of searching for a score solution can be generalized and
by doing so, further updates can occur smoothly. The recursive
backtracking now operates on both the sequence and section level.
Solving a sequence is delegated to an appropriate solver, producing a
solved sequence that has a uniform type, irrespective of the solver.